refactor(notifications): pushNotificationService 800줄 → internal 모듈 + 파사드
한도(800줄) 도달 파일을 notificationService.ts 파사드 패턴으로 분할:
state(공유홀더+lazy ref 접근자)·types·constants·permission·token·handlers 6 모듈 +
동일 경로 thin 파사드. public export 면(6타입·DEFAULT_CHANNELS·19 함수·getToken 별칭·
pushNotificationService 객체·default) 100% 보존 → 소비처(index.ts·hooks·settings·bridge·
test) 무수정.

linchpin: 모듈 간 가변상태는 단일 pushState 홀더 + getNotifications() 접근자로 공유
(exported let 미신뢰). bare Notifications 참조는 const Notifications = getNotifications()로
치환, null 가드·!단언 보존. 파사드 상단 부수효과 없음(순환 import 안전).

검증: tsc 0 · eslint 0 · 86/86(push+transient+notificationService+useFCMTokenManager) ·
Red-Green(getCurrentToken 공유단절 시 '등록 토큰 반환' 실패 확인 후 복원). 각 internal
파일 <800줄. 원본 export 면 diff 일치.
